chore(cli): organize lazy platform API factory
