+++ apps/cli/src/legacy/commands/bootstrap/bootstrap.layers.ts
@@ -1,12 +1,9 @@
-import { Effect, Layer } from "effect";
+import { Layer } from "effect";
 
 import { legacyCredentialsLayer } from "../../auth/legacy-credentials.layer.ts";
 import { legacyHttpClientLayer } from "../../auth/legacy-http-debug.layer.ts";
+import { legacyPlatformApiFactoryFromApiLayer } from "../../auth/legacy-platform-api-factory.layer.ts";
 import { legacyPlatformApiLayer } from "../../auth/legacy-platform-api.layer.ts";
-import {
-  LegacyPlatformApi,
-  LegacyPlatformApiFactory,
-} from "../../auth/legacy-platform-api.service.ts";
 import { legacyCliConfigLayer } from "../../config/legacy-cli-config.layer.ts";
 import { legacyProjectRefLayer } from "../../config/legacy-project-ref.layer.ts";
 import { legacyDebugLoggerLayer } from "../../shared/legacy-debug-logger.layer.ts";
@@ -46,25 +43,15 @@ const platformApi = legacyPlatformApiLayer.pipe(
   Layer.provide(httpClient),
   Layer.provide(debugLogger),
 );
-const platformApiFactory = Layer.effect(
-  LegacyPlatformApiFactory,
-  LegacyPlatformApi.pipe(
-    Effect.map((api) =>
-      LegacyPlatformApiFactory.of({
-        make: Effect.succeed(api),
-      }),
-    ),
-  ),
-);
-const platformApiFactoryStack = platformApiFactory.pipe(Layer.provide(platformApi));
+const platformApiFactory = legacyPlatformApiFactoryFromApiLayer.pipe(Layer.provide(platformApi));
 
 export const legacyBootstrapRuntimeLayer = Layer.mergeAll(
   platformApi,
-  platformApiFactoryStack,
+  platformApiFactory,
   httpClient,
   credentials,
   cliConfig,
-  legacyProjectRefLayer.pipe(Layer.provide(platformApiFactoryStack), Layer.provide(cliConfig)),
+  legacyProjectRefLayer.pipe(Layer.provide(platformApiFactory), Layer.provide(cliConfig)),
   legacyLinkedProjectCacheLayer.pipe(
     Layer.provide(credentials),
     Layer.provide(cliConfig),
