+++ apps/cli/src/legacy/commands/gen/types/types.layers.ts
@@ -1,17 +1,13 @@
-import { FetchHttpClient } from "effect/unstable/http";
-import { Effect, FileSystem, Layer, Path } from "effect";
-import * as HttpClient from "effect/unstable/http/HttpClient";
+import { Layer } from "effect";
 
 import { legacyCredentialsLayer } from "../../../auth/legacy-credentials.layer.ts";
-import { LegacyCredentials } from "../../../auth/legacy-credentials.service.ts";
-import { legacyMakePlatformApi } from "../../../auth/legacy-platform-api.layer.ts";
-import { LegacyPlatformApiFactory } from "../../../auth/legacy-platform-api.service.ts";
+import { legacyPlatformApiFactoryLayer } from "../../../auth/legacy-platform-api-factory.layer.ts";
+import { LegacyPlatformApiFactory } from "../../../auth/legacy-platform-api-factory.service.ts";
 import { legacyCliConfigLayer } from "../../../config/legacy-cli-config.layer.ts";
 import { LegacyCliConfig } from "../../../config/legacy-cli-config.service.ts";
 import { legacyProjectRefLayer } from "../../../config/legacy-project-ref.layer.ts";
 import { LegacyProjectRefResolver } from "../../../config/legacy-project-ref.service.ts";
 import { legacyDebugLoggerLayer } from "../../../shared/legacy-debug-logger.layer.ts";
-import { LegacyDebugLogger } from "../../../shared/legacy-debug-logger.service.ts";
 import { legacyHttpClientLayer } from "../../../auth/legacy-http-debug.layer.ts";
 import { legacyLinkedProjectCacheLayer } from "../../../telemetry/legacy-linked-project-cache.layer.ts";
 import { LegacyLinkedProjectCache } from "../../../telemetry/legacy-linked-project-cache.service.ts";
@@ -19,8 +15,6 @@ import { legacyTelemetryStateLayer } from "../../../telemetry/legacy-telemetry-s
 import { LegacyTelemetryState } from "../../../telemetry/legacy-telemetry-state.service.ts";
 import { commandRuntimeLayer } from "../../../../shared/runtime/command-runtime.layer.ts";
 import { CommandRuntime } from "../../../../shared/runtime/command-runtime.service.ts";
-import { Analytics } from "../../../../shared/telemetry/analytics.service.ts";
-import { TelemetryRuntime } from "../../../../shared/telemetry/runtime.service.ts";
 
 /**
  * `gen types --local` and `--db-url` do not use the Management API, so this
@@ -35,44 +29,16 @@ export const legacyGenTypesRuntimeLayer = (() => {
     Layer.provide(cliConfig),
     Layer.provide(legacyDebugLoggerLayer),
   );
-  const platformApiFactory = Layer.effect(
-    LegacyPlatformApiFactory,
-    Effect.gen(function* () {
-      const analytics = yield* Analytics;
-      const cliConfigService = yield* LegacyCliConfig;
-      const credentialsService = yield* LegacyCredentials;
-      const debugLogger = yield* LegacyDebugLogger;
-      const fs = yield* FileSystem.FileSystem;
-      const path = yield* Path.Path;
-      const telemetryRuntime = yield* TelemetryRuntime;
-
-      return LegacyPlatformApiFactory.of({
-        make: legacyMakePlatformApi.pipe(
-          Effect.provideService(Analytics, analytics),
-          Effect.provideService(LegacyCliConfig, cliConfigService),
-          Effect.provideService(LegacyCredentials, credentialsService),
-          Effect.provideService(LegacyDebugLogger, debugLogger),
-          Effect.provideService(FileSystem.FileSystem, fs),
-          Effect.provideService(Path.Path, path),
-          Effect.provideService(TelemetryRuntime, telemetryRuntime),
-          Effect.provide(FetchHttpClient.layer),
-        ),
-      });
-    }),
-  );
-  const platformApiFactoryStack = platformApiFactory.pipe(
+  const platformApiFactory = legacyPlatformApiFactoryLayer.pipe(
     Layer.provide(credentials),
     Layer.provide(cliConfig),
     Layer.provide(legacyDebugLoggerLayer),
   );
 
   const built = Layer.mergeAll(
-    httpClient,
-    credentials,
     cliConfig,
-    legacyDebugLoggerLayer,
-    platformApiFactoryStack,
-    legacyProjectRefLayer.pipe(Layer.provide(platformApiFactoryStack), Layer.provide(cliConfig)),
+    platformApiFactory,
+    legacyProjectRefLayer.pipe(Layer.provide(platformApiFactory), Layer.provide(cliConfig)),
     legacyLinkedProjectCacheLayer.pipe(
       Layer.provide(credentials),
       Layer.provide(cliConfig),
@@ -80,7 +46,7 @@ export const legacyGenTypesRuntimeLayer = (() => {
     ),
     legacyTelemetryStateLayer,
     commandRuntimeLayer(["gen", "types"]),
-  ).pipe(Layer.provide(FetchHttpClient.layer));
+  );
 
   const _serviceCoverageCheck: Layer.Layer<LegacyGenTypesServices, unknown, unknown> = built;
   void _serviceCoverageCheck;
@@ -89,11 +55,8 @@ export const legacyGenTypesRuntimeLayer = (() => {
 })();
 
 type LegacyGenTypesServices =
-  | HttpClient.HttpClient
-  | LegacyCredentials
-  | LegacyCliConfig
-  | LegacyDebugLogger
   | LegacyPlatformApiFactory
+  | LegacyCliConfig
   | LegacyProjectRefResolver
   | LegacyLinkedProjectCache
   | LegacyTelemetryState
