+++ apps/cli/src/legacy/shared/legacy-management-api-runtime.layer.ts
@@ -1,14 +1,12 @@
-import { Effect, Layer } from "effect";
+import { Layer } from "effect";
 import type * as HttpClient from "effect/unstable/http/HttpClient";
 import { FetchHttpClient } from "effect/unstable/http";
 
 import { LegacyCredentials } from "../auth/legacy-credentials.service.ts";
 import { legacyCredentialsLayer } from "../auth/legacy-credentials.layer.ts";
 import { legacyHttpClientLayer } from "../auth/legacy-http-debug.layer.ts";
-import {
-  LegacyPlatformApi,
-  LegacyPlatformApiFactory,
-} from "../auth/legacy-platform-api.service.ts";
+import { legacyPlatformApiFactoryFromApiLayer } from "../auth/legacy-platform-api-factory.layer.ts";
+import { LegacyPlatformApi } from "../auth/legacy-platform-api.service.ts";
 import { legacyPlatformApiLayer } from "../auth/legacy-platform-api.layer.ts";
 import { LegacyCliConfig } from "../config/legacy-cli-config.service.ts";
 import { legacyCliConfigLayer } from "../config/legacy-cli-config.layer.ts";
@@ -68,24 +66,16 @@ export function legacyManagementApiRuntimeLayer(subcommand: ReadonlyArray<string
     Layer.provide(FetchHttpClient.layer),
     Layer.provide(legacyDebugLoggerLayer),
   );
-  const platformApiFactory = Layer.effect(
-    LegacyPlatformApiFactory,
-    LegacyPlatformApi.pipe(
-      Effect.map((api) =>
-        LegacyPlatformApiFactory.of({
-          make: Effect.succeed(api),
-        }),
-      ),
-    ),
+  const platformApiFactory = legacyPlatformApiFactoryFromApiLayer.pipe(
+    Layer.provide(platformApiStack),
   );
-  const platformApiFactoryStack = platformApiFactory.pipe(Layer.provide(platformApiStack));
   const built = Layer.mergeAll(
     platformApiStack,
-    platformApiFactoryStack,
+    platformApiFactory,
     httpClient,
     credentials,
     cliConfig,
-    legacyProjectRefLayer.pipe(Layer.provide(platformApiFactoryStack), Layer.provide(cliConfig)),
+    legacyProjectRefLayer.pipe(Layer.provide(platformApiFactory), Layer.provide(cliConfig)),
     legacyLinkedProjectCacheLayer.pipe(
       Layer.provide(credentials),
       Layer.provide(cliConfig),
@@ -130,7 +120,6 @@ export function legacyManagementApiRuntimeLayer(subcommand: ReadonlyArray<string
  */
 type LegacyManagementApiServices =
   | LegacyPlatformApi
-  | LegacyPlatformApiFactory
   | HttpClient.HttpClient
   | LegacyCredentials
   | LegacyCliConfig
